dashboard/requirements.txt
streamlit>=1.32,<2.0
plotly>=5.18,<6.0
networkx>=3.2,<4.0
pandas>=2.1,<3.0
numpy>=1.26,<2.0
matplotlib>=3.8,<4.0
yfinance>=0.2.36
feedparser>=6.0
python-dateutil>=2.8
